requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.11.1
async-timeout==5.0.1
attrs==25.4.0
babel==2.18.0
billiard==4.2.4
brotli==1.2.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cssselect2==0.8.0
diff-match-patch==20241021
Django==4.2.29
django-allauth==65.14.3
django-appconf==1.2.0
django-celery==3.1.17
django-cors-headers==4.9.0
django-crispy-forms==1.8.0
django-csp==4.0
django-filter==25.1
django-import-export==4.4.0
django-mathfilters==1.0.0
django-model-utils==5.0.0
django-money==3.5.4
django-notifications-hq==1.8.3
django-qr-code==4.0.1
django-select2==8.2.4
django-simple-history==3.10.1
django-tinymce==5.0.0
django-widget-tweaks==1.5.1
djangorestframework==3.16.1
docopt==0.6.2
drf-spectacular==0.29.0
exceptiongroup==1.3.1
fonttools==4.60.2
gunicorn==23.0.0
idna==3.11
importlib_metadata==8.7.1
inflection==0.5.1
jsonfield==3.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.2
num2words==0.5.14
packaging==26.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
py-moneyed==3.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.0.1
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
segno==1.6.6
simplejson==3.20.2
six==1.17.0
slick-reporting==0.1.3
sqlparse==0.5.5
swapper==1.4.0
tablib==3.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
weasyprint==66.0
webencodings==0.5.1
zipp==3.23.0
zopfli==0.2.3.post1
